Reword to be better reflect RFC 2119. (#73)
Section 7.4.3 used the capitalized verb "MAY" without the semantics imposed by
RFC 2119.  The offending phrase is simply describing a situation that can
arise in the real world, rather than a requirement of the specification.  The
requirement put forth in this specification is defined in the next sentence
using the “MUST” verb.