Add link to consistency guidelines in README (#134)
* Reword to be better reflect RFC 2119. (#73)

Section 7.4.3 used the capitalized verb "MAY" without the semantics imposed by
RFC 2119.  The offending phrase is simply describing a situation that can
arise in the real world, rather than a requirement of the specification.  The
requirement put forth in this specification is defined in the next sentence
using the “MUST” verb.

* Fix formatting of contibutor table (#81)

Probably caused by different markdown implementation in github.

* Add link to consistency fundamentals to README

This is a quality of life add to the README to link directly to the consistency fundamentals section of the guidelines.