Update Guidelines.md
Fix rendering of the group-version correspondence table in the example in `12.1.1 Group versioning`. The table is interpreted as code blocks because of the four-space at the beginning of each line. Removing the leading spaces is one way to correct it, but for readability in text, I corrected it with leading `|`s instead.